Move the placement contract alongside the suggestions it pins
The test and its fixture were filed next to PaymentsProviders because that is the
class they call. That mirrors the invocation surface rather than the subject: the
fixture is a transcription of PaymentsExtensionSuggestions::$country_extensions,
it replaces tests that lived beside that class, and the test body speaks the
suggestions vocabulary far more than the settings one.

Anyone changing a country's recommendations edits the suggestions source, so the
contract that constrains those edits belongs in the same neighbourhood.

The fixture is renamed to say what it holds. A fixtures directory usually
contains inputs, as the neighbouring RemoteFreeExtensions and email editor
fixtures do; this file is the assertion side, and the `expected-` prefix settles
that before anyone opens it.

The class keeps its own file rather than merging into
PaymentsExtensionSuggestionsTest. That one is a unit test built on a mocked
incentives provider with no store state, while this needs a container-wired
service, an administrator, an enabled gateway, and profiler manipulation.
Merging would impose all of that on forty unrelated unit tests.

No behaviour changes: 609 tests, same assertions, and the fixture path printed in
failure messages is updated to match.
